+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # CHANGELOG
 
+## 2.7.25 - 2015-03-26
+
+* `Aws\ElasticTranscoder` - Added support for job timing.
+* `Aws\Iam` - Adding NamedPolicy to GetAccountAuthorizationDetails.
+* `Aws\OpsWorks` - Adding BlockDeviceMapping support.
+
 ## 2.7.24 - 2015-03-24
 
 * `Aws\S3` - Added support for cross-region replication.